requirements.txt
fastapi
uvicorn
mysql-connector-python
pymongo
python-dotenv
